ulong is stored as numeric(20,0) in pg anyway; drop the casting code

+++ src/Assistant.Net/Data/Entities/GameStatEntity.cs
@@ -7,11 +7,11 @@ namespace Assistant.Net.Data.Entities;
 [PrimaryKey(nameof(GuildId), nameof(UserId), nameof(GameType))]
 public class GameStatEntity
 {
-    public decimal GuildId { get; set; }
+    public ulong GuildId { get; set; }
 
     [ForeignKey(nameof(GuildId))] public GuildEntity Guild { get; set; } = null!;
 
-    public decimal UserId { get; set; }
+    public ulong UserId { get; set; }
 
     [ForeignKey(nameof(UserId))] public UserEntity User { get; set; } = null!;
