+++ src/Assistant.Net/Services/Logging/VoiceLogger.cs
@@ -51,7 +51,7 @@ public class VoiceLogger(
 
             if (!logConfig.IsEnabled || logConfig.ChannelId == null) return;
 
-            var webhookClient = await webhookService.GetOrCreateWebhookClientAsync((ulong)logConfig.ChannelId.Value)
+            var webhookClient = await webhookService.GetOrCreateWebhookClientAsync(logConfig.ChannelId.Value)
                 .ConfigureAwait(false);
             if (webhookClient == null) return;
